Rename file using snake_case
